+++ src/components/ConfigGlobal/src/types.ts
@@ -1,4 +1,3 @@
-import { ElementPlusSize } from './elementPlus'
 export interface ConfigGlobalTypes {
   size?: ElementPlusSize
 }
