+++ src/components/Descriptions/src/types.ts
