--- src/types/elementPlus.d.ts
@@ -1,3 +1,0 @@
-export type ElementPlusSize = 'default' | 'small' | 'large'
-
-export type ElementPlusInfoType = 'success' | 'info' | 'warning' | 'danger'
