+++ src/types/infoTip.ts
