--- src/types/layout.d.ts
@@ -1,1 +1,0 @@
-export type LayoutType = 'classic' | 'topLeft' | 'top' | 'cutMenu'
