+++ src/types/qrcode.ts
